Time: 27 ms (12.16%), Space: 48.1 MB (23.92%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.java
@@ -12,7 +12,7 @@ class Solution {
         while(low <= high){
             int mid = (low + high) >>> 1;
 
-            long hours = 0;
+            int hours = 0;
             for(int pile : piles){
                 hours += Math.ceil((double) pile / mid);
             }
